Implemented purge option for log utility + refactored args process
