+++ src/functions/logs/helpers/process-args.js
@@ -1,13 +1,38 @@
-module.exports = async (args) => {
-  switch (args[0]) {
-    case '--purge':
-      console.log('purge')
-      break
-    case '--watch':
-      console.log('watch')
-      break
-    default:
-      const chalk = require('chalk')
-      throw new Error(`${args[0]} is not a supported option for ${chalk.italic.blue('logs')} utility. Available options are: [--purge] to purge log file(s) & [--watch] to watch a log file.`)
+module.exports = async (args, callback) => {
+  if (args.length === 0) {
+    args = (await prompt()).fileNames
   }
+  for (const arg of args) {
+    process(arg, callback)
+  }
+}
+
+function process(arg, callback) {
+  const fs = require('fs')
+  const path = require('path')
+  const chalk = require('chalk')
+  const file = `${arg}.log`
+  const filePath = path.join(appRoot, 'logs', file)
+  clog.info(`Looking for log file called ${chalk.italic.cyan(file)}...`)
+  if (fs.existsSync(filePath)) {
+    clog.success(`${chalk.italic.cyan(file)} log file found!`)
+    callback(file, filePath)
+  } else {
+    clog.error(`${chalk.italic.cyan(file)} log file does not exist!`)
+    console.log('\n')
+  }
+}
+
+async function prompt() {
+  const inquirer = require('inquirer')
+  const fs = require('fs')
+  const path = require('path')
+  return await inquirer.prompt([
+    {
+      type: 'checkbox',
+      name: 'fileNames',
+      message: 'Select log files to print:',
+      choices: fs.readdirSync(path.join(appRoot, 'logs'), {withFileTypes: true}).filter(file => file.isFile()).map(file => file.name.replace('.log', ''))
+    }
+  ])
 }
