+++ src/functions/logs/helpers/watch.js
@@ -1,3 +1,3 @@
 module.exports = async (args) => {
-  
+  console.log('watch')
 }
